Added option to copy password
The password is copied via Ctrl C from the password field if enabled
in the preferences dialog. Only enable it if you work on a secure
system, every one of your application will be able to access the data
while on the clipboard.

NOTE: Some applications including the Gnome terminal seems to have
problems with the data, however Firefox seems to work.